+++ frontend/src/components/Hazard/GMS/GmsViewer.js
@@ -1,7 +1,0 @@
-import React, { Fragment } from "react";
-
-const GmsViewer = () => {
-  return <Fragment>Coming Soon!</Fragment>;
-};
-
-export default GmsViewer;