+++ frontend/src/components/Hazard/SeismicHazard/HazardCurveSection.js
@@ -1,5 +1,7 @@
 import React, { useState, useContext, useEffect, Fragment } from "react";
-import { IMSelect } from "components/common";
+import IMSelect from "components/common/IMSelect";
+import GuideTooltip from "components/common/GuideTooltip";
+import * as CONSTANTS from "constants/Constants";
 import { v4 as uuidv4 } from "uuid";
 import { GlobalContext } from "context";
 
@@ -24,7 +26,10 @@ const HazardCurveSection = () => {
   return (
     <Fragment>
       <div className="form-group form-section-title">
-        <span>Hazard Curve</span>
+        Hazard Curve
+        <GuideTooltip
+          explanation={CONSTANTS.TOOLTIP_MESSAGES["HAZARD_HAZARD"]}
+        />
       </div>
       <div className="custom-form-group">
         <IMSelect title="Intensity Measure" setIM={setSelectedIM} />