+++ frontend/src/components/Hazard/SeismicHazard/HazardViewerDisaggregation.js
@@ -229,7 +229,7 @@ const HazadViewerDisaggregation = () => {
           {disaggComputeClick === null && (
             <GuideMessage
               header={CONSTANTS.DISAGGREGATION}
-              body={CONSTANTS.DISAGGREGATION_WARNING_MSG_PLOT}
+              body={CONSTANTS.DISAGGREGATION_GUIDE_MSG_PLOT}
               instruction={CONSTANTS.DISAGGREGATION_INSTRUCTION_PLOT}
             />
           )}
@@ -261,7 +261,7 @@ const HazadViewerDisaggregation = () => {
           {disaggComputeClick === null && (
             <GuideMessage
               header={CONSTANTS.DISAGGREGATION}
-              body={CONSTANTS.DISAGGREGATION_WARNING_MSG_PLOT}
+              body={CONSTANTS.DISAGGREGATION_GUIDE_MSG_PLOT}
               instruction={CONSTANTS.DISAGGREGATION_INSTRUCTION_PLOT}
             />
           )}
@@ -293,7 +293,7 @@ const HazadViewerDisaggregation = () => {
           {disaggComputeClick === null && (
             <GuideMessage
               header={CONSTANTS.DISAGGREGATION}
-              body={CONSTANTS.DISAGGREGATION_WARNING_MSG_TABLE}
+              body={CONSTANTS.DISAGGREGATION_GUIDE_MSG_TABLE}
               instruction={CONSTANTS.DISAGGREGATION_INSTRUCTION_TABLE}
             />
           )}