+++ frontend/src/components/Hazard/SeismicHazard/HazardViewerHazardCurve.js
@@ -176,7 +176,7 @@ const HazardViewerHazardCurve = () => {
           {hazardCurveComputeClick === null && (
             <GuideMessage
               header={CONSTANTS.HAZARD_CURVE}
-              body={CONSTANTS.HAZARD_CURVE_WARNING_MSG}
+              body={CONSTANTS.HAZARD_CURVE_GUIDE_MSG}
               instruction={CONSTANTS.HAZARD_CURVE_INSTRUCTION}
             />
           )}
@@ -217,7 +217,7 @@ const HazardViewerHazardCurve = () => {
           {hazardCurveComputeClick === null && (
             <GuideMessage
               header={CONSTANTS.HAZARD_CURVE}
-              body={CONSTANTS.HAZARD_CURVE_WARNING_MSG}
+              body={CONSTANTS.HAZARD_CURVE_GUIDE_MSG}
               instruction={CONSTANTS.HAZARD_CURVE_INSTRUCTION}
             />
           )}