+++ frontend/src/components/Hazard/SeismicHazard/NZS1170Section.js
@@ -4,6 +4,7 @@ import TextField from "@material-ui/core/TextField";
 import Select from "react-select";
 import * as CONSTANTS from "constants/Constants";
 import { useAuth0 } from "components/common/ReactAuth0SPA";
+import GuideTooltip from "components/common/GuideTooltip";
 
 import "assets/style/NZS1170Section.css";
 import { handleErrors, disableScrollOnNumInput } from "utils/Utils";
@@ -60,15 +61,18 @@ const NZS1170Section = () => {
     Based on them, create an array to be used in react-select.
   */
   useEffect(() => {
-    const tempArr = [];
+    if (Object.entries(soilClass).length > 0) {
+      // Due to value.replaceAll can't be applied on null, check the object is not empty
+      const tempArr = [];
 
-    for (const [key, value] of Object.entries(soilClass)) {
-      tempArr.push({
-        value: key,
-        label: `${key} - ${value.replaceAll("_", " ")}`,
-      });
+      for (const [key, value] of Object.entries(soilClass)) {
+        tempArr.push({
+          value: key,
+          label: `${key} - ${value.replaceAll("_", " ")}`,
+        });
+      }
+      setLocalSoilClasses(tempArr);
     }
-    setLocalSoilClasses(tempArr);
   }, [soilClass]);
 
   /*
@@ -358,7 +362,11 @@ const NZS1170Section = () => {
     <Fragment>
       <form autoComplete="off" onSubmit={(e) => e.preventDefault()}>
         <div className="form-group form-section-title">
-          <span>NZS1170.5</span>
+          NZS1170.5
+          <GuideTooltip
+            explanation={CONSTANTS.TOOLTIP_MESSAGES["HAZARD_NZCODE"]}
+            hyperlink={CONSTANTS.TOOLTIP_URL["HAZARD_NZCODE"]}
+          />
         </div>
 
         <div className="form-group">