+++ frontend/src/components/Project/SeismicHazard/HazardViewerDisaggregation.js
@@ -233,7 +233,7 @@ const HazadViewerDisaggregation = () => {
           {projectDisaggGetClick === null && (
             <GuideMessage
               header={CONSTANTS.DISAGGREGATION}
-              body={CONSTANTS.DISAGGREGATION_WARNING_MSG_PLOT}
+              body={CONSTANTS.DISAGGREGATION_GUIDE_MSG_PLOT}
               instruction={CONSTANTS.PROJECT_DISAGG_INSTRUCTION_PLOT}
             />
           )}
@@ -265,7 +265,7 @@ const HazadViewerDisaggregation = () => {
           {projectDisaggGetClick === null && (
             <GuideMessage
               header={CONSTANTS.DISAGGREGATION}
-              body={CONSTANTS.DISAGGREGATION_WARNING_MSG_PLOT}
+              body={CONSTANTS.DISAGGREGATION_GUIDE_MSG_PLOT}
               instruction={CONSTANTS.PROJECT_DISAGG_INSTRUCTION_PLOT}
             />
           )}
@@ -297,7 +297,7 @@ const HazadViewerDisaggregation = () => {
           {projectDisaggGetClick === null && (
             <GuideMessage
               header={CONSTANTS.DISAGGREGATION}
-              body={CONSTANTS.DISAGGREGATION_WARNING_MSG_TABLE}
+              body={CONSTANTS.DISAGGREGATION_GUIDE_MSG_TABLE}
               instruction={CONSTANTS.PROJECT_DISAGG_INSTRUCTION_TABLE}
             />
           )}