+++ frontend/src/components/Project/SeismicHazard/HazardViewerHazardCurve.js
@@ -139,7 +139,7 @@ const HazardViewerHazardCurve = () => {
           {projectHazardCurveGetClick === null && (
             <GuideMessage
               header={CONSTANTS.HAZARD_CURVE}
-              body={CONSTANTS.HAZARD_CURVE_WARNING_MSG}
+              body={CONSTANTS.HAZARD_CURVE_GUIDE_MSG}
               instruction={CONSTANTS.PROJECT_HAZARD_CURVE_INSTRUCTION}
             />
           )}
@@ -180,7 +180,7 @@ const HazardViewerHazardCurve = () => {
           {projectHazardCurveGetClick === null && (
             <GuideMessage
               header={CONSTANTS.HAZARD_CURVE}
-              body={CONSTANTS.HAZARD_CURVE_WARNING_MSG}
+              body={CONSTANTS.HAZARD_CURVE_GUIDE_MSG}
               instruction={CONSTANTS.PROJECT_HAZARD_CURVE_INSTRUCTION}
             />
           )}