+++ frontend/src/constants/Constants.js
@@ -24,6 +24,10 @@ export const CORE_API_ROUTE_GMS_DEFAULT_IM_WEIGHTS =
   "coreAPI/gms/default_im_weights";
 export const CORE_API_ROUTE_GMS_DEFAULT_CAUSAL_PARAMS =
   "coreAPI/gms/default_causal_params";
+export const CORE_API_ROUTE_GMS_GET_AVAILABLE_GMS =
+  "coreAPI/gms/ensemble_gms/ims";
+export const CORE_API_ROUTE_GMS_GET_GM_DATASETS =
+  "coreAPI/gms/ensemble_gms/datasets";
 
 // This endpoint will eventually replace when we implement DB properly
 // As this function reads from Available_Project table (A bridge table between User and Project)
@@ -88,6 +92,7 @@ export const SITE_SELECTION_VS30_TITLE = "VS30";
 
 /* 
   Guide Messages
+  Site Selection - Regional & VS30
 */
 export const SITE_SELECTION_VS30_MSG =
   "Please do the following steps to see an image.";
@@ -127,18 +132,18 @@ export const DEFAULT_ANNUAL_PROB =
   process.env.REACT_APP_DEFAULT_ANNUAL_EXCEEDANCE_RATE || "";
 
 /* 
-  Hazard Viewer
   Guide messages
+  Seismic Hazard - Hazard Viewer
 */
-export const HAZARD_CURVE_WARNING_MSG =
+export const HAZARD_CURVE_GUIDE_MSG =
   "Please do the following steps to see plots.";
 export const HAZARD_CURVE_INSTRUCTION = [
   "Choose the Intensity Measure first.",
   "Click the compute button in the 'Hazard Curve' section to see plots.",
 ];
-export const DISAGGREGATION_WARNING_MSG_PLOT =
+export const DISAGGREGATION_GUIDE_MSG_PLOT =
   "Please do the following steps to see plots.";
-export const DISAGGREGATION_WARNING_MSG_TABLE =
+export const DISAGGREGATION_GUIDE_MSG_TABLE =
   "Please do the following steps to see the contribution table.";
 export const DISAGGREGATION_INSTRUCTION_PLOT = [
   "Choose the Intensity Measure first.",
@@ -159,10 +164,24 @@ export const UNIFORM_HAZARD_SPECTRUM_INSTRUCTION = [
   "Click Compute button to see plots.",
 ];
 
+/* 
+  Guide messages
+  - GMS
+*/
+export const GMS_VIEWER_GUIDE_MSG =
+  "Please do the following steps to see plots.";
+export const GMS_VIEWER_GUIDE_INSTRUCTION = [
+  "Select IM from IM Type.",
+  "Choose IM Level or Exceedance Rate then put a value.",
+  "Select IM Vector(s).",
+  "Put Number of Ground Motions.",
+  "Click the Compute button.",
+];
+
 /*
   Project Tabs
 */
-export const PROJECT_SITE_SELECTION_WARNING_MSG =
+export const PROJECT_SITE_SELECTION_GUIDE_MSG =
   "Please do the following steps to see images.";
 export const PROJECT_SITE_SELECTION_INSTRUCTION = [
   "Choose the Project ID.",
@@ -212,7 +231,7 @@ export const ERROR_SET_DIFF_CODE = {
 };
 
 /*
-  Hazard Curve Plots
+  react-plotly.js configuration
 */
 export const PLOT_MARGIN = {
   l: 60,
@@ -238,4 +257,41 @@ export const PLOT_CONFIG = {
     "autoScale2d",
   ],
   doubleClick: "autosize",
+};
+
+/*
+  Constant Tooltips Message
+*/
+export const TOOLTIP_MESSAGES = {
+  SITE_SELECTION_LOCATION:
+    "Enter the location (Lat, Lon) of the site of interest",
+  SITE_SELECTION_SITE_CONDITION:
+    "Enter the site parameters, such as the 30m-averaged shear-wave velocity. Default values of these parameters are estimated based on the selected location. You can either use this or manually override it with a specified value.",
+  HAZARD_HAZARD:
+    "Select the intensity measure of interest to compute the seismic hazard curve.",
+  HAZARD_DISAGG:
+    "Specify the annual exceedance rate (inverse of return period) to compute the disaggregation distribution. The adopted intensity measure is that specified in the Intensity Measure tab.",
+  HAZARD_UHS:
+    "Specify one or more annual exceedance rates (inverse of return period) to compute the Uniform Hazard Spectrum (UHS) for. Specified rates are displayed, which can be subsequently removed.",
+  HAZARD_NZCODE:
+    "Select the parameters of NZS1170.5:2004 to compare with the site-specific hazard results. Based on the location and Vs30 values assigned, these parameters have been estimated, but can be manually over-ridden.",
+  PROJECT_SITE_SELECTION_PROJECT_ID:
+    "Select the Project title from the list of available alternatives.",
+  PROJECT_SITE_SELECTION_LOCATION:
+    "Select the site/location of interest for this Project.",
+  PROJECT_SITE_SELECTION_VS30:
+    "Select the site parameters of interest for this site of interest, such as the 30m-averaged shear-wave velocity.",
+  PROJECT_HAZARD:
+    "Select the intensity measure of interest to compute the seismic hazard curve.",
+  PROJECT_DISAGG:
+    "Select the annual exceedance rate (inverse of return period) to compute the disaggregation distribution. The adopted intensity measure is that specified in the Intensity Measure tab.",
+  PROJECT_UHS:
+    "Select one or more annual exceedance rates (inverse of return period) to compute the Uniform Hazard Spectrum (UHS) for. Selected rates are displayed, which can be subsequently removed.",
+};
+
+/*
+  Hyperlinks for Tooltips
+*/
+export const TOOLTIP_URL = {
+  HAZARD_NZCODE: "https://google.com",
 };